move more callbacks into callbacks module
