+++ crates/core/src/client/config.rs
@@ -2,11 +2,7 @@ use std::{collections::HashMap, sync::Arc};
 
 use phoenix_channels_client::JSON;
 
-use crate::{
-    callbacks::{SecurePersistentStore, *},
-    dom::DocumentChangeHandler,
-    live_socket::LiveFile,
-};
+use crate::{callbacks::*, live_socket::LiveFile};
 
 #[derive(uniffi::Enum, Debug, Clone, Default, Copy)]
 pub enum LogLevel {
