+++ crates/core/src/live_socket/tests/mod.rs
@@ -1,8 +1,9 @@
 use std::{sync::Arc, time::Duration};
 
 use super::*;
-use crate::dom::{
-    ChangeType, ControlFlow, DocumentChangeHandler, LiveChannelStatus, NodeData, NodeRef,
+use crate::{
+    callbacks::*,
+    dom::{NodeData, NodeRef},
 };
 mod error;
 mod navigation;
